Update: Finished the basic functionality
Managed to make an object out of the booking class and also structured the page how it would look like. It will still be improved but for now its okay. Also, changed the data for the images and used urls from the internet instead of referencing them locally. And most importantly, managed to link the selected image to the booking page using it's id and getting it's data from the database using that id. I am proud to have come so far with minimal help.